Add edge to edge (e2e) tests for git submodules
Edge to edge tests in benchbuild should avoid relying on the cli
and any real projects to test our API.

This commits adds 2 e2e tests, when we use Git and GitSubmodules:

1. Check, if a projects emits any versions, when Submodules are involved.
2. Check, if we can execution a full benchbuild run. Without touching any
   real compilation/execution commands.
